打卡单个列表判断更新

+++ src/components/user/mine.vue
@@ -1,10 +1,10 @@
 <!--
  * @Author: your name
  * @Date: 2021-10-13 10:51:39
- * @LastEditTime: 2021-11-04 08:56:20
- * @LastEditors: Please set LastEditors
+ * @LastEditTime: 2022-03-26 22:16:04
+ * @LastEditors: 王朋坤
  * @Description: In User Settings Edit
- * @FilePath: /graduation-project-master/src/components/mine/mine.vue
+ * @FilePath: /graduation-project-master/src/components/user/mine.vue
 -->
 
 <template>
@@ -13,7 +13,7 @@
     <div class="mine-form-user-information">
       <img src="@/assets/img/user/1.jpg" alt="" />
       <div class="user-main">
-        <div>{{ $store.state.User.login.name }}</div>
+        <div>{{ "王朋坤" }}</div>
         <div>{{ "2018248*****" }}</div>
       </div>
       <div class="user-jump"  :style="{'background-image': `url(${require('@/assets/font/arrow.svg')})`}"></div>
@@ -40,13 +40,33 @@
 
 <script>
 import "@/assets/style/common/mobile-form.scss"
+import { Dialog } from 'vant';
 export default {
+  components: {
+    [Dialog.Component.name]: Dialog.Component.name,
+  },
   methods: {
     loginOut() {
-      this.$router.push("/user/login");
-      window.sessionStorage.clear("loginData");
-      this.$notify({ type: "primary", message: "注销登陆成功"});
-    }
+    this.$dialog.confirm({
+      title: '是否退出登陆'
+       
+    })
+      .then(() => {
+        // on confirm
+        this.$router.push("/user/login");
+        window.sessionStorage.clear("loginData");
+        this.$notify({ type: "primary", message: "注销登陆成功"});
+      })
+      .catch(() => {
+        // on cancel
+        this.$toast("取消成功");
+      });
+
+    },
+    // components: {
+    //   [Dialog.name]: Dialog
+    // }
+
   },
   mounted() {
     // this.$notify({type: "primary", message: "欢迎来到个人中心"});
@@ -59,6 +79,11 @@ export default {
   overflow-y: auto;
   width: 100%;
   height: 100%;
+}
+
+.van-dialog .van-dialog--default {
+  border-radius: 8px !important;
+  // border-radius: 16px;
 }
 
 </style>
